feat: statically defined ECCVM/Translator VK data (#13395)
Introduces static methods for initializing the verification keys for
ECCVM/Translator. This is possible since the VK commitments for these
VMs are a function only of the respective fixed circuit size constants.
The ECCVM/Translator VKs are now default constructed with the correct
values which allows removal of a lot of explicit handling of them
elsewhere.

Note: there's more simplification that could come out of this work but
its beyond the scope of the primary purpose of this PR which was to
introduce simple methods for obtaining the fixed VK data.

Closes https://github.com/AztecProtocol/barretenberg/issues/798