chore: delete honk_recursion for building ACIR (#13664)
Closes https://github.com/AztecProtocol/barretenberg/issues/1013 in the
context for building ACIR.

When we updated verify_proof_with_type to no longer default to plonk
proofs, we forgot to also delete some hacky code wrt using the
honk_recursion flag for figuring out which recursive verifier to use.
Since we always receive the proof type through the actual constraint,
there's no need to use the honk_recursion flag to pick this out, so it
can be completely removed for this context.