fix: more worker fixes (#13759)
Of course webpack wasn't going to cooperate.

* Ignore webpack's web-worker loader and do it ourselves.
* Fix the web build since it doesn't make sense for it to include TS
stuff
* Ensure all the downstream projects work with new bundling

Cleaned up flakes:

* `bb` browser test seems to have a nice regexp for the flake, the
"catchall" version only hid real errors
* Only timeouts on boxes may be flakes, the locator.toBeVisible() hides
real errors

Also noticed the THREAD_MODEL in `acir_tests` was not being honored, we
were always running in multithreaded mode (which is what makes sense for
proving/verifying), so got rid of it. Also, the requisites for single
threaded bb.js are strictly a subset of the ones for multithreading, so
we're fine.

Edit: 

Had to split the test due to running into this webkit bug:
https://bugs.webkit.org/show_bug.cgi?id=245346

---------

Co-authored-by: Nicolás Venturo <nicolas.venturo@gmail.com>