chore: start translator logic at an even index (#13985)
We want to be able to add random data in the ultra op queue (at the
beginning and end) to make the merge protocol zk without affecting the
logic of translator or the version of the op queue used by eccvm. All
wires in translator circuit builder start with a 0 to enable shifting.
But having the builder add data in all wires, including the ones
containing op queue data, breaks the ability of the Goblin verifier to
compare the full table commitments in the last merge against the
corresponding translator witness polynomials commitments.

To solve this we want to add the 0 row (plus random rows eventually) via
the ultra op queue logic, but each ultra op populates two positions in
the translator wires. Prior to this work, the translator relations were
implemented to expect the main logic to start at an odd index (i.e. the
first ultra op resides at index 1 and 2). Preserving this would have
meant we need to implement a special branch of logic in the ultra op
queue that only populates 1 row with data rather than 2 in the ultra
tables. This PR swaps what happens at even and odd indexes to facilitate
adding 0 and random rows via the existing op queue logic by making
translator logic start at an even index (so currently at index 2).